Barreled enemy, enemy deal damage

+++ Assets/Scripts/Character/Enemy/EnemyDamage.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class EnemyDamage : MonoBehaviour, IDamageable
 {
     // Made by Lukas and Anton A 2026-03-06
-    [field: SerializeField] public int MaxHealth { get; set; }
-    [HideInInspector] public int Health { get; set; }
+    [field: SerializeField] public float MaxHealth { get; set; }
+    [HideInInspector] public float Health { get; set; }
     [HideInInspector] public bool CanTakeDamage { get; set; } = true;
     private ITriggerable onDeath;
 
@@ -15,7 +15,7 @@ public class EnemyDamage : MonoBehaviour, IDamageable
     float damageCooldownTimer = 1;
     [SerializeField] float damageCooldown = 1;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (CanTakeDamage && Health > 0)
         {
