+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -4,11 +4,11 @@ public interface IDamageable
     // Made by Lukas and Anton B 2026-03-06
     // Updated by Lukas and Jonatan 2026-03-16
 
-    public int MaxHealth { get;}
-    public int Health { get; }
+    public float MaxHealth { get;}
+    public float Health { get; }
     public bool CanTakeDamage { get; }
 
     //IEnumerator DamageCoolDown(int damageDelay); // Uncomment if you want to use a cooldown for taking damage, but currently not used in the project
-    public void TakeDamage(int damage);
+    public void TakeDamage(float damage);
     public void Death();
 }
